+++ runtime/bin/platform_linux.cc
@@ -110,11 +110,6 @@ const char* Platform::ResolveExecutablePath() {
 }
 
 
-bool Platform::AnsiSupported() {
-  return true;
-}
-
-
 void Platform::Exit(int exit_code) {
   exit(exit_code);
 }